requirements.txt
requests==2.32.3
beautifulsoup4==4.13.3
langchain==0.3.23
sentence-transformers>=3.0.0
chromadb>=0.5.0
transformers>=4.40.0
torch>=2.1.0
accelerate>=0.28.0
peft>=0.10.0
bitsandbytes>=0.43.0
django>=5.0
djangorestframework>=3.15.0
psycopg2-binary>=2.9.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
python-dotenv>=1.0.0
